Introduces the Scorer component
