+++ consensus/core/src/authority_node.rs
@@ -29,6 +29,7 @@ use crate::{
     metrics::initialise_metrics,
     network::{NetworkClient as _, NetworkManager, tonic_network::TonicManager},
     round_prober::{RoundProber, RoundProberHandle},
+    scorer::Scorer,
     storage::rocksdb_store::RocksDBStore,
     subscriber::Subscriber,
     synchronizer::{Synchronizer, SynchronizerHandle},
@@ -61,6 +62,7 @@ impl ConsensusAuthority {
         // running. It's mostly useful to make decisions on whether amnesia recovery should
         // run or not. When `boot_counter` is 0, then `ConsensusAuthority` will initiate
         // the process of amnesia recovery if that's enabled in the parameters.
+        scorer: Arc<Scorer>,
         boot_counter: u64,
     ) -> Self {
         match network_type {
@@ -75,6 +77,7 @@ impl ConsensusAuthority {
                     transaction_verifier,
                     commit_consumer,
                     registry,
+                    scorer,
                     boot_counter,
                 )
                 .await;
@@ -158,6 +161,7 @@ where
         transaction_verifier: Arc<dyn TransactionVerifier>,
         commit_consumer: CommitConsumer,
         registry: Registry,
+        scorer: Arc<Scorer>,
         boot_counter: u64,
     ) -> Self {
         assert!(
@@ -178,13 +182,13 @@ where
         );
         info!("Consensus parameters: {:?}", parameters);
         info!("Consensus committee: {:?}", committee);
-        let committee_size = committee.size();
         let context = Arc::new(Context::new(
             own_index,
             committee,
             parameters,
             protocol_config,
-            initialise_metrics(registry, committee_size),
+            initialise_metrics(registry),
+            scorer,
             Arc::new(Clock::new()),
         ));
         let start_time = Instant::now();
@@ -434,6 +438,7 @@ mod tests {
     use crate::{
         CommittedSubDag,
         block::{BlockAPI as _, GENESIS_ROUND},
+        scorer,
         transaction::NoopTransactionVerifier,
     };
 
@@ -458,6 +463,7 @@ mod tests {
 
         let (sender, _receiver) = unbounded_channel("consensus_output");
         let commit_consumer = CommitConsumer::new(sender, 0);
+        let scorer = Arc::new(Scorer::new_dummy_for_tests(committee.size()));
 
         let authority = ConsensusAuthority::start(
             network_type,
@@ -470,6 +476,7 @@ mod tests {
             Arc::new(txn_verifier),
             commit_consumer,
             registry,
+            scorer,
             0,
         )
         .await;
@@ -852,6 +859,7 @@ mod tests {
 
         let (sender, receiver) = unbounded_channel("consensus_output");
         let commit_consumer = CommitConsumer::new(sender, 0);
+        let scorer = Arc::new(scorer::Scorer::new_dummy_for_tests(committee.size()));
 
         let authority = ConsensusAuthority::start(
             network_type,
@@ -864,6 +872,7 @@ mod tests {
             Arc::new(txn_verifier),
             commit_consumer,
             registry,
+            scorer,
             boot_counter,
         )
         .await;
