+++ consensus/core/src/metrics.rs
@@ -2,16 +2,8 @@
 // Modifications Copyright (c) 2024 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-use std::{
-    collections::BTreeSet,
-    sync::{
-        Arc,
-        atomic::{AtomicU64, Ordering},
-    },
-};
+use std::sync::Arc;
 
-use consensus_config::AuthorityIndex;
-use itertools::izip;
 use prometheus::{
     Histogram, HistogramVec, IntCounter, IntCounterVec, IntGauge, IntGaugeVec, Registry,
     exponential_buckets, register_histogram_vec_with_registry, register_histogram_with_registry,
@@ -19,7 +11,7 @@ use prometheus::{
     register_int_gauge_vec_with_registry, register_int_gauge_with_registry,
 };
 
-use crate::{BlockRef, context::Context, error::ConsensusError, network::metrics::NetworkMetrics};
+use crate::network::metrics::NetworkMetrics;
 
 // starts from 1μs, 50μs, 100μs...
 const FINE_GRAINED_LATENCY_SEC_BUCKETS: &[f64] = &[
@@ -88,340 +80,7 @@ const SIZE_BUCKETS: &[f64] = &[
 pub(crate) struct Metrics {
     pub(crate) node_metrics: NodeMetrics,
     pub(crate) network_metrics: NetworkMetrics,
-    pub(crate) scoring_metrics: ValidatorScoringMetrics,
 }
-
-enum MetricType {
-    Cached,
-    Uncached,
-}
-
-impl Metrics {
-    pub(crate) fn update_scoring_metrics_on_eviction(
-        &self,
-        authority_index: AuthorityIndex,
-        hostname: &str,
-        recent_refs: &BTreeSet<BlockRef>,
-        eviction_round: u32,
-        last_eviction_round: u32,
-        threshold_clock_round: u32,
-    ) -> Option<StoredScoringMetricsU64> {
-        // threshold_clock_round should be always at least 1.
-        // Analogously, authority_index should be a valid index.
-        if threshold_clock_round == 0
-            || authority_index.value() >= self.scoring_metrics.cached.len()
-        {
-            return None;
-        }
-
-        // Get the blocks rounds that were not evicted.
-        let cached_block_rounds = recent_refs
-            .iter()
-            .map(|block| block.round)
-            .filter(|&round| round > eviction_round && round < threshold_clock_round)
-            .collect::<Vec<u32>>();
-
-        // Update metrics according to the blocks from rounds still in cache.
-        let (cached_equivocations, missing_blocks_in_cached_rounds) =
-            calculate_scoring_metrics_for_range(
-                cached_block_rounds,
-                eviction_round + 1,
-                threshold_clock_round.saturating_sub(1),
-            );
-
-        self.update_missing_blocks_and_equivocations(
-            missing_blocks_in_cached_rounds,
-            cached_equivocations,
-            hostname,
-            authority_index,
-            MetricType::Cached,
-        );
-
-        // If no eviction happened, we do not update the metrics on storage.
-        if eviction_round == last_eviction_round {
-            return None;
-        }
-
-        // Get the evicted blocks rounds.
-        let evicted_block_rounds = recent_refs
-            .iter()
-            .map(|block| block.round)
-            .filter(|&round| round <= eviction_round)
-            .collect::<Vec<u32>>();
-
-        // Update metrics according to the blocks from evicted rounds.
-        let (evicted_equivocations, missing_blocks_in_evicted_rounds) =
-            calculate_scoring_metrics_for_range(
-                evicted_block_rounds,
-                last_eviction_round + 1,
-                eviction_round,
-            );
-
-        self.update_missing_blocks_and_equivocations(
-            missing_blocks_in_evicted_rounds,
-            evicted_equivocations,
-            hostname,
-            authority_index,
-            MetricType::Uncached,
-        );
-
-        // Update score
-        self.update_authority_score(authority_index, hostname);
-
-        Some(StoredScoringMetricsU64 {
-            faulty_blocks_provable: self.scoring_metrics.uncached[authority_index]
-                .faulty_blocks_provable
-                .load(Ordering::Relaxed),
-            faulty_blocks_unprovable: self.scoring_metrics.uncached[authority_index]
-                .faulty_blocks_unprovable
-                .load(Ordering::Relaxed),
-            equivocations: self.scoring_metrics.uncached[authority_index]
-                .equivocations
-                .load(Ordering::Relaxed),
-            missing_proposals: self.scoring_metrics.uncached[authority_index]
-                .missing_proposals
-                .load(Ordering::Relaxed),
-        })
-    }
-
-    pub(crate) fn initialize_scoring_metrics(
-        &self,
-        recovered_scoring_metrics: Vec<(AuthorityIndex, StoredScoringMetricsU64)>,
-        blocks_in_cache_by_authority: &Vec<BTreeSet<BlockRef>>,
-        threshold_clock_round: u32,
-        eviction_rounds: &Vec<u32>,
-        context: Arc<Context>,
-    ) {
-        let hostnames = context
-            .committee
-            .authorities()
-            .map(|(_, x)| x.hostname.as_str())
-            .collect::<Vec<_>>();
-
-        for ((authority_index, metrics), hostname, blocks_in_cache, &eviction_round) in izip!(
-            recovered_scoring_metrics,
-            hostnames,
-            blocks_in_cache_by_authority,
-            eviction_rounds
-        ) {
-            // Initialize the uncached scoring metrics according to
-            // recovered_scoring_metrics
-            let StoredScoringMetricsU64 {
-                faulty_blocks_provable,
-                faulty_blocks_unprovable,
-                equivocations,
-                missing_proposals,
-            } = metrics;
-            self.initialize_faulty_blocks_metrics(
-                faulty_blocks_provable,
-                faulty_blocks_unprovable,
-                hostname,
-                authority_index,
-            );
-            self.update_missing_blocks_and_equivocations(
-                missing_proposals,
-                equivocations,
-                hostname,
-                authority_index,
-                MetricType::Uncached,
-            );
-
-            // Initialize the cached scoring metrics according to blocks_in_cache.
-            let block_rounds_in_cache = blocks_in_cache
-                .iter()
-                .map(|block_ref| block_ref.round)
-                .collect();
-            let (cached_equivocations, missing_blocks_in_cached_rounds) =
-                calculate_scoring_metrics_for_range(
-                    block_rounds_in_cache,
-                    eviction_round + 1,
-                    threshold_clock_round - 1,
-                );
-            self.update_missing_blocks_and_equivocations(
-                missing_blocks_in_cached_rounds,
-                cached_equivocations,
-                hostname,
-                authority_index,
-                MetricType::Cached,
-            );
-            // Initiatize score
-            self.update_authority_score(authority_index, hostname);
-        }
-    }
-
-    // Auxiliary function to update scoring metrics relative to missing blocks
-    // and equivocations. The `authority` parameter should be a valid index,
-    // otherwise the function will panic. This check is not performed here, as
-    // it is assumed that the caller has already checked it.
-    fn update_missing_blocks_and_equivocations(
-        &self,
-        missing_blocks: u64,
-        equivocations: u64,
-        hostname: &str,
-        authority: AuthorityIndex,
-        metric_type: MetricType,
-    ) {
-        match metric_type {
-            MetricType::Cached => {
-                self.scoring_metrics.cached[authority]
-                    .equivocations
-                    .store(equivocations, Ordering::Relaxed);
-                self.scoring_metrics.cached[authority]
-                    .missing_proposals
-                    .store(missing_blocks, Ordering::Relaxed);
-                self.node_metrics
-                    .equivocations_in_cache_by_authority
-                    .with_label_values(&[hostname])
-                    .set(equivocations as i64);
-                self.node_metrics
-                    .missing_proposals_in_cache_by_authority
-                    .with_label_values(&[hostname])
-                    .set(missing_blocks as i64);
-            }
-
-            MetricType::Uncached => {
-                self.scoring_metrics.uncached[authority]
-                    .equivocations
-                    .fetch_add(equivocations, Ordering::Relaxed);
-                self.scoring_metrics.uncached[authority]
-                    .missing_proposals
-                    .fetch_add(missing_blocks, Ordering::Relaxed);
-                self.node_metrics
-                    .uncached_equivocations_by_authority
-                    .with_label_values(&[hostname])
-                    .inc_by(equivocations);
-                self.node_metrics
-                    .uncached_missing_proposals_by_authority
-                    .with_label_values(&[hostname])
-                    .inc_by(missing_blocks);
-            }
-        }
-    }
-
-    // Auxiliary function used to update scores. The `authority` parameter should be
-    // a valid index, otherwise the function will panic. This check is not
-    // performed here, as it is assumed that the caller has already checked it.
-    fn update_authority_score(&self, authority: AuthorityIndex, hostname: &str) {
-        let (faulty_blocks_provable, faulty_blocks_unprovable, equivocations, missing_proposals) = (
-            self.scoring_metrics.uncached[authority]
-                .faulty_blocks_provable
-                .load(Ordering::Relaxed),
-            self.scoring_metrics.uncached[authority]
-                .faulty_blocks_unprovable
-                .load(Ordering::Relaxed),
-            self.scoring_metrics.uncached[authority]
-                .equivocations
-                .load(Ordering::Relaxed),
-            self.scoring_metrics.uncached[authority]
-                .missing_proposals
-                .load(Ordering::Relaxed),
-        );
-
-        // We provisionally use the formula below to calculate the score, but changes
-        // to this function are already expected. The hardcoded parameters (which are
-        // also still provisional) represent:
-        // - No tolerance to any provably faulty misbehaviour. If any is detected, the
-        //   score will be 0.
-        // - If no provably faulty blocks are detected, 50% of the score is guaranteed.
-        // - If no unprovably faulty blocks are detected, an additional 12.5% of the
-        //   score is guaranteed.
-        // - If no missing proposals are detected, an additional 37.5% of the score is
-        //   guaranteed.
-        // - The maximum achievable score is u32::MAX.
-
-        if faulty_blocks_provable > 0 || equivocations > 0 {
-            self.scoring_metrics.score[authority].store(0);
-            self.node_metrics
-                .score_by_authority
-                .with_label_values(&[hostname])
-                .set(0i64);
-        } else {
-            let score = (2 << 31) - 1
-                + (3 * (2 << 29) / (missing_proposals.saturating_add(1))
-                    + (2 << 29) / (faulty_blocks_unprovable.saturating_add(1)));
-            self.scoring_metrics.score[authority].store(score);
-            self.node_metrics
-                .score_by_authority
-                .with_label_values(&[hostname])
-                .set(score as i64);
-        }
-    }
-
-    // Auxiliary function to initialize scoring metrics relative to faulty blocks.
-    // The `authority` parameter should be a valid index, otherwise the function
-    // will panic. This check is not performed here, as it is assumed that the
-    // caller has already checked it.
-    fn initialize_faulty_blocks_metrics(
-        &self,
-        faulty_blocks_provable: u64,
-        faulty_blocks_unprovable: u64,
-        hostname: &str,
-        authority_index: AuthorityIndex,
-    ) {
-        self.node_metrics
-            .faulty_blocks_provable_by_authority
-            .with_label_values(&[hostname, "loaded from storage", "loaded from storage"])
-            .inc_by(faulty_blocks_provable);
-        self.node_metrics
-            .faulty_blocks_unprovable_by_authority
-            .with_label_values(&[hostname, "loaded from storage", "loaded from storage"])
-            .inc_by(faulty_blocks_unprovable);
-        self.scoring_metrics.uncached[authority_index]
-            .faulty_blocks_provable
-            .store(faulty_blocks_provable, Ordering::Relaxed);
-        self.scoring_metrics.uncached[authority_index]
-            .faulty_blocks_unprovable
-            .store(faulty_blocks_unprovable, Ordering::Relaxed);
-    }
-
-    pub(crate) fn update_scoring_metrics_on_block_receival(
-        &self,
-        authority_index: AuthorityIndex,
-        hostname: &str,
-        error: ConsensusError,
-        source: &str,
-    ) {
-        // authority_index will be always a valid index. However, this method will
-        // panic if authority_index >= committee_size. We run this check only to avoid
-        // this panic.
-        if authority_index.value() >= self.scoring_metrics.cached.len() {
-            return;
-        }
-
-        if should_update_provable_metrics(&error, source) {
-            self.scoring_metrics.uncached[authority_index]
-                .faulty_blocks_provable
-                .fetch_add(1, Ordering::Relaxed);
-            self.node_metrics
-                .faulty_blocks_provable_by_authority
-                .with_label_values(&[hostname, source, error.name()])
-                .inc();
-        } else if should_update_unprovable_metrics(&error, source) {
-            self.scoring_metrics.uncached[authority_index]
-                .faulty_blocks_unprovable
-                .fetch_add(1, Ordering::Relaxed);
-            self.node_metrics
-                .faulty_blocks_unprovable_by_authority
-                .with_label_values(&[hostname, source, error.name()])
-                .inc();
-        } else {
-            return;
-        }
-    }
-}
-
-pub(crate) struct Score(AtomicU64);
-
-impl Score {
-    pub(crate) fn new() -> Self {
-        Self(AtomicU64::new(u64::MAX))
-    }
-
-    pub(crate) fn store(&self, value: u64) {
-        self.0.store(value, Ordering::Relaxed);
-    }
-}
-
 pub(crate) struct NodeMetrics {
     pub(crate) block_commit_latency: Histogram,
     pub(crate) proposed_blocks: IntCounterVec,
@@ -1189,1339 +848,16 @@ impl NodeMetrics {
     }
 }
 
-pub(crate) struct ValidatorScoringMetrics {
-    pub(crate) uncached: Vec<UncachedScoringMetrics>,
-    pub(crate) cached: Vec<CachedScoringMetrics>,
-    pub(crate) score: Vec<Score>,
-}
-
-impl ValidatorScoringMetrics {
-    pub(crate) fn new(committee_size: usize) -> Self {
-        let uncached = (0..committee_size)
-            .map(|_| UncachedScoringMetrics::new())
-            .collect();
-        let cached = (0..committee_size)
-            .map(|_| CachedScoringMetrics::new())
-            .collect();
-        let score = (0..committee_size).map(|_| Score::new()).collect();
-        Self {
-            uncached,
-            cached,
-            score,
-        }
-    }
-}
-
-#[derive(Debug)]
-pub(crate) struct UncachedScoringMetrics {
-    // Counts the number of times that a faulty block signed by the validator was already verified
-    // in the epoch.
-    pub(crate) faulty_blocks_provable: AtomicU64,
-    // Counts the number of times that a faulty block not signed by the validator was already
-    // verified in the epoch.
-    pub(crate) faulty_blocks_unprovable: AtomicU64,
-    // Counts the number of equivocations that were already evicted from cache in the epoch.
-    pub(crate) equivocations: AtomicU64,
-    // Counts the number of blocks that the validator failed to propose, or that the node did not
-    // receive, from the rounds already evicted from cache in the epoch.
-    pub(crate) missing_proposals: AtomicU64,
-}
-
-impl UncachedScoringMetrics {
-    pub(crate) fn new() -> Self {
-        Self {
-            faulty_blocks_provable: AtomicU64::new(0),
-            faulty_blocks_unprovable: AtomicU64::new(0),
-            equivocations: AtomicU64::new(0),
-            missing_proposals: AtomicU64::new(0),
-        }
-    }
-}
-
-pub(crate) struct CachedScoringMetrics {
-    // Counts the number of equivocations in cache, below the threshold clock round.
-    pub(crate) equivocations: AtomicU64,
-    // Counts the number of blocks that the validator failed to propose, or that the node did not
-    // receive yet, from the rounds stored in cache and below the threshold clock round.
-    pub(crate) missing_proposals: AtomicU64,
-}
-
-impl CachedScoringMetrics {
-    pub(crate) fn new() -> Self {
-        Self {
-            equivocations: AtomicU64::new(0),
-            missing_proposals: AtomicU64::new(0),
-        }
-    }
-}
-
-// This struct is used in storage. It holds the same data as
-// `UncachedScoringMetrics`, but uses `u64` instead of `AtomicU64`.
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
-pub(crate) struct StoredScoringMetricsU64 {
-    pub(crate) faulty_blocks_provable: u64,
-    pub(crate) faulty_blocks_unprovable: u64,
-    pub(crate) equivocations: u64,
-    pub(crate) missing_proposals: u64,
-}
-
-pub(crate) fn initialise_metrics(registry: Registry, committee_size: usize) -> Arc<Metrics> {
+pub(crate) fn initialise_metrics(registry: Registry) -> Arc<Metrics> {
     let node_metrics = NodeMetrics::new(&registry);
     let network_metrics = NetworkMetrics::new(&registry);
-    let scoring_metrics = ValidatorScoringMetrics::new(committee_size);
     Arc::new(Metrics {
         node_metrics,
         network_metrics,
-        scoring_metrics,
     })
 }
 
-// Given the set of blocks issued by an authority in rounds in the inclusive
-// range [start, end], this function calculates and returns the number of
-// equivocations and missing blocks in that range . The function should receive
-// the vector with the rounds of such blocks and the range start and end points.
-fn calculate_scoring_metrics_for_range(
-    mut block_rounds: Vec<u32>,
-    start: u32,
-    end: u32,
-) -> (u64, u64) {
-    // Filter out rounds that are not in the range [start, end].
-    block_rounds.retain(|&round| round >= start && round <= end);
-    let number_of_blocks = block_rounds.len();
-    block_rounds.dedup();
-    let unique_block_rounds = block_rounds.len();
-    // We use saturating_sub to avoid unexpected underflows, but the subtractions
-    // below should never result in negative values by construction:
-    // 1) unique_block_rounds <= number_of_blocks
-    // 2) end - start + 1 >= unique_block_rounds
-    let number_of_equivocations = number_of_blocks.saturating_sub(unique_block_rounds) as u64;
-    let number_of_missing_blocks =
-        (end + 1).saturating_sub(start + unique_block_rounds as u32) as u64;
-
-    (number_of_equivocations, number_of_missing_blocks)
-}
-
-fn should_update_provable_metrics(error: &ConsensusError, source: &str) -> bool {
-    if source == "handle_send_block"
-        && (is_from_signed_block_verification(error)
-            || matches!(
-                error,
-                ConsensusError::BlockRejected { .. }
-                //| ConsensusError::MalformedAncestorBlock { .. }
-                ))
-    {
-        return true;
-    }
-    false
-}
-
-fn should_update_unprovable_metrics(error: &ConsensusError, source: &str) -> bool {
-    if source == "handle_send_block" {
-        return is_from_unsigned_block_verification(error)
-            || matches!(
-                error,
-                ConsensusError::MalformedBlock { .. } | ConsensusError::UnexpectedAuthority { .. }
-            );
-    } else if source == "fetch_once" {
-        return is_from_commit_syncer(error);
-    } else if source == "process_fetched_blocks" {
-        return is_from_unsigned_block_verification(error)
-            || is_from_signed_block_verification(error)
-            || matches!(error, ConsensusError::MalformedBlock { .. });
-    }
-    false
-}
-
-fn is_from_unsigned_block_verification(err: &ConsensusError) -> bool {
-    matches!(
-        err,
-        ConsensusError::WrongEpoch { .. }
-            | ConsensusError::UnexpectedGenesisBlock
-            | ConsensusError::InvalidAuthorityIndex { .. }
-            | ConsensusError::SerializationFailure { .. }
-            | ConsensusError::MalformedSignature { .. }
-            | ConsensusError::SignatureVerificationFailure { .. }
-    )
-}
-
-fn is_from_signed_block_verification(err: &ConsensusError) -> bool {
-    matches!(
-        err,
-        ConsensusError::TooManyAncestors { .. }
-            | ConsensusError::InsufficientParentStakes { .. }
-            | ConsensusError::InvalidAuthorityIndex { .. }
-            | ConsensusError::InvalidAncestorPosition { .. }
-            | ConsensusError::InvalidAncestorRound { .. }
-            | ConsensusError::InvalidGenesisAncestor { .. }
-            | ConsensusError::DuplicatedAncestorsAuthority { .. }
-            | ConsensusError::TransactionTooLarge { .. }
-            | ConsensusError::TooManyTransactions { .. }
-            | ConsensusError::TooManyTransactionBytes { .. }
-            | ConsensusError::InvalidTransaction { .. }
-    )
-}
-
-fn is_from_commit_syncer(err: &ConsensusError) -> bool {
-    matches!(
-        err,
-        ConsensusError::MalformedCommit { .. }
-            | ConsensusError::UnexpectedStartCommit { .. }
-            | ConsensusError::UnexpectedCommitSequence { .. }
-            | ConsensusError::NoCommitReceived { .. }
-            | ConsensusError::MalformedBlock { .. }
-            | ConsensusError::NotEnoughCommitVotes { .. }
-            | ConsensusError::UnexpectedNumberOfBlocksFetched { .. }
-            | ConsensusError::UnexpectedBlockForCommit { .. }
-    ) || is_from_unsigned_block_verification(err)
-        || is_from_signed_block_verification(err)
-}
-
 #[cfg(test)]
-pub(crate) fn test_metrics(committee_size: usize) -> Arc<Metrics> {
-    initialise_metrics(Registry::new(), committee_size)
-}
-
-#[cfg(test)]
-mod tests {
-    use std::{
-        collections::BTreeSet,
-        sync::{Arc, atomic::Ordering},
-        time::Duration,
-        vec,
-    };
-
-    use async_trait::async_trait;
-    use bytes::Bytes;
-    use consensus_config::{AuthorityIndex, NetworkKeyPair, ProtocolKeyPair};
-    use parking_lot::RwLock;
-    use prometheus::Registry;
-    use tokio::sync::broadcast;
-
-    use crate::{
-        Round, TransactionVerifier, ValidationError,
-        authority_service::{AuthorityService, tests::FakeCoreThreadDispatcher},
-        block::{BlockDigest, BlockRef, VerifiedBlock},
-        block_verifier::SignedBlockVerifier,
-        commit::CommitRange,
-        commit_vote_monitor::CommitVoteMonitor,
-        context::Context,
-        dag_state::DagState,
-        error::ConsensusResult,
-        metrics::{
-            ConsensusError, StoredScoringMetricsU64, ValidatorScoringMetrics, initialise_metrics,
-        },
-        network::{BlockStream, NetworkClient},
-        storage::mem_store::MemStore,
-        synchronizer::Synchronizer,
-        test_dag_builder::DagBuilder,
-    };
-    struct TxnSizeVerifier {}
-
-    impl TransactionVerifier for TxnSizeVerifier {
-        fn verify_batch(&self, _transactions: &[&[u8]]) -> Result<(), ValidationError> {
-            unimplemented!("Unimplemented")
-        }
-    }
-
-    #[derive(Default)]
-    struct FakeNetworkClient {}
-
-    #[async_trait]
-    impl NetworkClient for FakeNetworkClient {
-        const SUPPORT_STREAMING: bool = false;
-
-        async fn send_block(
-            &self,
-            _peer: AuthorityIndex,
-            _block: &VerifiedBlock,
-            _timeout: Duration,
-        ) -> ConsensusResult<()> {
-            unimplemented!("Unimplemented")
-        }
-
-        async fn subscribe_blocks(
-            &self,
-            _peer: AuthorityIndex,
-            _last_received: Round,
-            _timeout: Duration,
-        ) -> ConsensusResult<BlockStream> {
-            unimplemented!("Unimplemented")
-        }
-
-        async fn fetch_blocks(
-            &self,
-            _peer: AuthorityIndex,
-            _block_refs: Vec<BlockRef>,
-            _highest_accepted_rounds: Vec<Round>,
-            _timeout: Duration,
-        ) -> ConsensusResult<Vec<Bytes>> {
-            unimplemented!("Unimplemented")
-        }
-
-        async fn fetch_commits(
-            &self,
-            _peer: AuthorityIndex,
-            _commit_range: CommitRange,
-            _timeout: Duration,
-        ) -> ConsensusResult<(Vec<Bytes>, Vec<Bytes>)> {
-            unimplemented!("Unimplemented")
-        }
-
-        async fn fetch_latest_blocks(
-            &self,
-            _peer: AuthorityIndex,
-            _authorities: Vec<AuthorityIndex>,
-            _timeout: Duration,
-        ) -> ConsensusResult<Vec<Bytes>> {
-            unimplemented!("Unimplemented")
-        }
-
-        async fn get_latest_rounds(
-            &self,
-            _peer: AuthorityIndex,
-            _timeout: Duration,
-        ) -> ConsensusResult<(Vec<Round>, Vec<Round>)> {
-            unimplemented!("Unimplemented")
-        }
-    }
-
-    // Creates a new authority service for scoring metrics testing purposes.
-    fn new_authority_service_for_metrics_tests(
-        committee_size: usize,
-    ) -> (
-        Vec<(NetworkKeyPair, ProtocolKeyPair)>,
-        Arc<Context>,
-        Arc<FakeCoreThreadDispatcher>,
-        Arc<AuthorityService<FakeCoreThreadDispatcher>>,
-    ) {
-        let (context, keys) = Context::new_for_test(committee_size);
-        let context = Arc::new(context);
-        let block_verifier = Arc::new(SignedBlockVerifier::new(
-            context.clone(),
-            Arc::new(TxnSizeVerifier {}),
-        ));
-        let commit_vote_monitor = Arc::new(CommitVoteMonitor::new(context.clone()));
-        let core_dispatcher = Arc::new(FakeCoreThreadDispatcher::new());
-        let (_tx_block_broadcast, rx_block_broadcast) = broadcast::channel(100);
-        let network_client = Arc::new(FakeNetworkClient::default());
-        let store = Arc::new(MemStore::new());
-        let dag_state = Arc::new(RwLock::new(DagState::new(context.clone(), store.clone())));
-        let synchronizer = Synchronizer::start(
-            network_client,
-            context.clone(),
-            core_dispatcher.clone(),
-            commit_vote_monitor.clone(),
-            block_verifier.clone(),
-            dag_state.clone(),
-            false,
-        );
-        let authority_service = Arc::new(AuthorityService::new(
-            context.clone(),
-            block_verifier,
-            commit_vote_monitor,
-            synchronizer,
-            core_dispatcher.clone(),
-            rx_block_broadcast,
-            dag_state,
-            store,
-        ));
-        (keys, context, core_dispatcher, authority_service)
-    }
-
-    impl ValidatorScoringMetrics {
-        pub(crate) fn uncached_missing_proposals_by_authority(&self) -> Vec<u64> {
-            self.uncached
-                .iter()
-                .map(|metrics| metrics.missing_proposals.load(Ordering::Relaxed))
-                .collect()
-        }
-
-        pub(crate) fn equivocations_in_cache_by_authority(&self) -> Vec<u64> {
-            self.cached
-                .iter()
-                .map(|metrics| metrics.equivocations.load(Ordering::Relaxed))
-                .collect()
-        }
-
-        pub(crate) fn missing_proposals_in_cache_by_authority(&self) -> Vec<u64> {
-            self.cached
-                .iter()
-                .map(|metrics| metrics.missing_proposals.load(Ordering::Relaxed))
-                .collect()
-        }
-
-        pub(crate) fn uncached_equivocations_by_authority(&self) -> Vec<u64> {
-            self.uncached
-                .iter()
-                .map(|metrics| metrics.equivocations.load(Ordering::Relaxed))
-                .collect()
-        }
-
-        pub(crate) fn faulty_blocks_provable_by_authority(&self) -> Vec<u64> {
-            self.uncached
-                .iter()
-                .map(|metrics| metrics.faulty_blocks_provable.load(Ordering::Relaxed))
-                .collect()
-        }
-
-        pub(crate) fn faulty_blocks_unprovable_by_authority(&self) -> Vec<u64> {
-            self.uncached
-                .iter()
-                .map(|metrics| metrics.faulty_blocks_unprovable.load(Ordering::Relaxed))
-                .collect()
-        }
-    }
-
-    #[test]
-    fn test_update_scoring_metrics_on_eviction_edge_cases() {
-        let metrics = initialise_metrics(Registry::new(), 4);
-        let authority_index = AuthorityIndex::new_for_test(0);
-        let hostname = "test_host";
-        let recent_refs_by_authority = BTreeSet::new();
-
-        // Test different unexpected combinations of eviction_round, last_evicted_round,
-        // and threshold_clock_round. Since recent_refs_by_authority is empty, the
-        // function should never panic or return more than zero equivocations.
-        // Each of the cases below have a small explanation of why they are unexpected
-        // and why they are supposed to return what they return.
-
-        // Unexpected because: threshold_clock_round = last_evicted_round means that a
-        // round with blocks from less than 2f+1 stake was evicted.
-        // Return: None, because nothing is currently being evicted.
-        let last_evicted_round = 5;
-        let eviction_round = 5;
-        let threshold_clock_round = 5;
-        let stored_metrics = metrics.update_scoring_metrics_on_eviction(
-            authority_index,
-            hostname,
-            &recent_refs_by_authority,
-            eviction_round,
-            last_evicted_round,
-            threshold_clock_round,
-        );
-        assert!(stored_metrics.is_none());
-
-        // Unexpected because: threshold_clock_round = 0 means that genesis is missing.
-        // Return: None, because nothing is currently being evicted.
-        let last_evicted_round = 0;
-        let eviction_round = 0;
-        let threshold_clock_round = 0;
-        let stored_metrics = metrics.update_scoring_metrics_on_eviction(
-            authority_index,
-            hostname,
-            &recent_refs_by_authority,
-            eviction_round,
-            last_evicted_round,
-            threshold_clock_round,
-        );
-        assert!(stored_metrics.is_none());
-
-        // Unexpected because: threshold_clock_round < eviction_round means that a
-        // round with blocks from less than 2f+1 stake in being evicted.
-        // Return: 3 missing proposals, from rounds 1 to 3 (eviction_round).
-        let last_evicted_round = 0;
-        let eviction_round = 3;
-        let threshold_clock_round = 2;
-        let stored_metrics = metrics.update_scoring_metrics_on_eviction(
-            authority_index,
-            hostname,
-            &recent_refs_by_authority,
-            eviction_round,
-            last_evicted_round,
-            threshold_clock_round,
-        );
-        assert!(matches!(
-            stored_metrics,
-            Some(StoredScoringMetricsU64 {
-                faulty_blocks_provable: 0,
-                faulty_blocks_unprovable: 0,
-                equivocations: 0,
-                missing_proposals: 3
-            })
-        ));
-
-        // Unexpected because: eviction_round < last_evicted_round means that blocks
-        // below or in last_evicted_round were accepted.
-        // Return: metrics won't be updated here, so it should return the same as in the
-        // last step.
-        let last_evicted_round = 1;
-        let eviction_round = 0;
-        let threshold_clock_round = 2;
-        let stored_metrics = metrics.update_scoring_metrics_on_eviction(
-            authority_index,
-            hostname,
-            &recent_refs_by_authority,
-            eviction_round,
-            last_evicted_round,
-            threshold_clock_round,
-        );
-        assert!(matches!(
-            stored_metrics,
-            Some(StoredScoringMetricsU64 {
-                faulty_blocks_provable: 0,
-                faulty_blocks_unprovable: 0,
-                equivocations: 0,
-                missing_proposals: 3
-            })
-        ));
-
-        // Unexpected because: threshold_clock_round < eviction_round <
-        // last_evicted_round and threshold_clock_round. Return: metrics won't
-        // be updated here, so it should return the same as in the last step.
-        let last_evicted_round = 2;
-        let eviction_round = 0;
-        let threshold_clock_round = 1;
-        let stored_metrics = metrics.update_scoring_metrics_on_eviction(
-            authority_index,
-            hostname,
-            &recent_refs_by_authority,
-            eviction_round,
-            last_evicted_round,
-            threshold_clock_round,
-        );
-        assert!(matches!(
-            stored_metrics,
-            Some(StoredScoringMetricsU64 {
-                faulty_blocks_provable: 0,
-                faulty_blocks_unprovable: 0,
-                equivocations: 0,
-                missing_proposals: 3
-            })
-        ));
-
-        // Unexpected because: threshold_clock_round < last_evicted_round means that a
-        // round with blocks from less than 2f+1 stake was evicted.
-        // Return: None, because nothing is currently being evicted.
-        let last_evicted_round = 1;
-        let eviction_round = 2;
-        let threshold_clock_round = 0;
-        let stored_metrics = metrics.update_scoring_metrics_on_eviction(
-            authority_index,
-            hostname,
-            &recent_refs_by_authority,
-            eviction_round,
-            last_evicted_round,
-            threshold_clock_round,
-        );
-        assert!(stored_metrics.is_none());
-
-        let last_evicted_round = 2;
-        let eviction_round = 1;
-        let threshold_clock_round = 0;
-        let stored_metrics = metrics.update_scoring_metrics_on_eviction(
-            authority_index,
-            hostname,
-            &recent_refs_by_authority,
-            eviction_round,
-            last_evicted_round,
-            threshold_clock_round,
-        );
-        assert!(stored_metrics.is_none());
-
-        // The function should not panic if the authority index is out of
-        // bounds.
-        // Unexpected because: threshold_clock_round = last_evicted_round means that a
-        // round with blocks from less than 2f+1 stake was evicted.
-        // Return: None, because nothing is currently being evicted.
-        let out_of_bounds_authority_index = AuthorityIndex::new_for_test(4);
-        let last_evicted_round = 1;
-        let eviction_round = 2;
-        let threshold_clock_round = 3;
-        let stored_metrics = metrics.update_scoring_metrics_on_eviction(
-            out_of_bounds_authority_index,
-            hostname,
-            &recent_refs_by_authority,
-            eviction_round,
-            last_evicted_round,
-            threshold_clock_round,
-        );
-        assert!(stored_metrics.is_none());
-    }
-
-    #[tokio::test]
-    async fn test_metrics_flush_and_recovery_gc_enabled() {
-        telemetry_subscribers::init_for_testing();
-
-        const GC_DEPTH: u32 = 3;
-        const CACHED_ROUNDS: u32 = 4;
-
-        let committee_size = 4;
-        let (mut context, _) = Context::new_for_test(committee_size);
-
-        context.parameters.dag_state_cached_rounds = CACHED_ROUNDS;
-        context
-            .protocol_config
-            .set_consensus_gc_depth_for_testing(GC_DEPTH);
-        context
-            .protocol_config
-            .set_consensus_linearize_subdag_v2_for_testing(true);
-
-        let context = Arc::new(context);
-        let metrics = &context.metrics.scoring_metrics;
-
-        let store = Arc::new(MemStore::new());
-        let mut dag_state = DagState::new(context.clone(), store.clone());
-
-        // Initialize the DAG builder with 20 layers. Blocks in the DAG will reference
-        // all blocks from the previous round.
-        // - Rounds 1 to 5 will have unique blocks from all authorities.
-        // - Rounds 6 to 8 will have unique blocks from all authorities, except 0, who
-        //   will not propose any block.
-        // - Rounds 9 to 10 will have unique blocks from all authorities.
-        // - Rounds 11 to 20 will have unique blocks from all authorities, except:
-        //      - Authority 1, who will produce 1 equivocating blocks at round 11 (i.e.,
-        //        1+1 blocks)
-        //      - Authority 2, who will produce 2 equivocating blocks at round 13 (i.e.,
-        //        1+2 blocks)
-        let mut dag_builder = DagBuilder::new(context.clone());
-        dag_builder.layers(1..=5).build();
-        dag_builder
-            .layers(6..=8)
-            .authorities(vec![AuthorityIndex::new_for_test(0)])
-            .skip_block()
-            .build();
-        dag_builder.layers(9..=10).build();
-        dag_builder
-            .layers(11..=11)
-            .authorities(vec![AuthorityIndex::new_for_test(1)])
-            .equivocate(1)
-            .build();
-        dag_builder.layers(12..=12).build();
-        dag_builder
-            .layers(13..=13)
-            .authorities(vec![AuthorityIndex::new_for_test(2)])
-            .equivocate(2)
-            .build();
-        dag_builder.layers(14..=20).build();
-
-        let mut commits = dag_builder
-            .get_sub_dag_and_commits(1..=20)
-            .into_iter()
-            .map(|(_subdag, commit)| commit)
-            .collect::<Vec<_>>();
-
-        // Add the blocks and commits from first 10 rounds to the dag state. Since
-        // authority 0 skipped a leader round, we use the 9 first items of the commits
-        // vector
-        let mut temp_commits = commits.split_off(9);
-        dag_state.accept_blocks(dag_builder.blocks(1..=10));
-        for commit in commits.clone() {
-            dag_state.add_commit(commit);
-        }
-
-        // Checks that metrics are still all zeroed, since even though we accepted
-        // blocks to the dag state, the metrics updates are done when the dag state is
-        // flushed.
-        assert_eq!(
-            [
-                metrics.uncached_equivocations_by_authority(),
-                metrics.uncached_missing_proposals_by_authority(),
-                metrics.equivocations_in_cache_by_authority(),
-                metrics.missing_proposals_in_cache_by_authority()
-            ],
-            [
-                vec![0; committee_size],
-                vec![0; committee_size],
-                vec![0; committee_size],
-                vec![0; committee_size]
-            ]
-        );
-
-        // Flush the dag state
-        dag_state.flush();
-
-        // Check that metrics were updated correctly after flushing.
-        //
-        // Equivocations:
-        // - We only accepted blocks from rounds <= 10, thus, no equivocations were
-        //   accepted yet. Equivocations metrics, then, should be still all zeroed.
-        //
-        // Missing proposals:
-        // - The last_commit_round should be 10, so gc_round should be 6. The eviction
-        //   round, then, should be 6 for all authorities.
-        // - The threshold_clock_round should be 11, since we already accepted all
-        //   blocks from epoch 10.
-        // - Then, finally, we should have counted:
-        //      - 1 uncached missing proposal for authority 0;
-        //      - 2 missing proposal in cache for authority 0;
-        //      - 0 missing proposals for authorities 1, 2, and 3.
-        assert_eq!(
-            [
-                metrics.uncached_equivocations_by_authority(),
-                metrics.uncached_missing_proposals_by_authority(),
-                metrics.equivocations_in_cache_by_authority(),
-                metrics.missing_proposals_in_cache_by_authority()
-            ],
-            [
-                vec![0; committee_size],
-                vec![1, 0, 0, 0],
-                vec![0; committee_size],
-                vec![2, 0, 0, 0],
-            ]
-        );
-
-        // Clear and check all metrics
-        metrics.uncached[0]
-            .missing_proposals
-            .store(0, Ordering::Relaxed);
-        metrics.cached[0]
-            .missing_proposals
-            .store(0, Ordering::Relaxed);
-        assert_eq!(
-            [
-                metrics.uncached_equivocations_by_authority(),
-                metrics.uncached_missing_proposals_by_authority(),
-                metrics.equivocations_in_cache_by_authority(),
-                metrics.missing_proposals_in_cache_by_authority()
-            ],
-            [
-                vec![0; committee_size],
-                vec![0; committee_size],
-                vec![0; committee_size],
-                vec![0; committee_size]
-            ]
-        );
-
-        // Destroy and recover dag state from storage.
-        drop(dag_state);
-        let mut dag_state = DagState::new(context.clone(), store.clone());
-
-        // Metrics should have been initialized as before the recovery.
-        assert_eq!(
-            [
-                metrics.uncached_equivocations_by_authority(),
-                metrics.uncached_missing_proposals_by_authority(),
-                metrics.equivocations_in_cache_by_authority(),
-                metrics.missing_proposals_in_cache_by_authority()
-            ],
-            [
-                vec![0; committee_size],
-                vec![1, 0, 0, 0],
-                vec![0; committee_size],
-                vec![2, 0, 0, 0],
-            ]
-        );
-
-        // Add blocks and commits from rounds 11 and 12 to the dag state.
-        let second_temp_commits = temp_commits.split_off(2);
-        dag_state.accept_blocks(dag_builder.blocks(11..=12));
-        for commit in temp_commits.clone() {
-            dag_state.add_commit(commit);
-        }
-
-        // Flush the dag state
-        dag_state.flush();
-
-        // Check that metrics were updated correctly after flushing.
-        //
-        // Missing proposals:
-        // - The last_commit_round should be 12, so gc_round should be 8. The eviction
-        //   round, then, should be 8 for all authorities. Then, we should have counted:
-        //      - 3 missing proposal in cache for authority 0;
-        //      - 0 missing proposals for authorities 1, 2, and 3.
-        //
-        // Equivocations:
-        // - We only removed from cache blocks from rounds <= 8, thus, no equivocations
-        //   should be uncached. Then, we should have counted:
-        //      - 0 uncached equivocations;
-        //      - 1 equivocation in cache for authority 1;
-        //      - 0 equivocations in cache for authorities 0, 2 and 3;
-        //
-
-        assert_eq!(
-            [
-                metrics.uncached_equivocations_by_authority(),
-                metrics.uncached_missing_proposals_by_authority(),
-                metrics.equivocations_in_cache_by_authority(),
-                metrics.missing_proposals_in_cache_by_authority()
-            ],
-            [
-                vec![0; committee_size],
-                vec![3, 0, 0, 0],
-                vec![0, 1, 0, 0],
-                vec![0; committee_size],
-            ]
-        );
-
-        // Accept all the rest of blocks and commits.
-        dag_state.accept_blocks(dag_builder.blocks(13..=20));
-        for commit in second_temp_commits.clone() {
-            dag_state.add_commit(commit);
-        }
-
-        // Clear and check all metrics
-        metrics.uncached[0]
-            .missing_proposals
-            .store(0, Ordering::Relaxed);
-        metrics.cached[1].equivocations.store(0, Ordering::Relaxed);
-
-        assert_eq!(
-            [
-                metrics.uncached_equivocations_by_authority(),
-                metrics.uncached_missing_proposals_by_authority(),
-                metrics.equivocations_in_cache_by_authority(),
-                metrics.missing_proposals_in_cache_by_authority()
-            ],
-            [
-                vec![0; committee_size],
-                vec![0; committee_size],
-                vec![0; committee_size],
-                vec![0; committee_size]
-            ]
-        );
-
-        // Destroy and recover dag state from storage.
-        drop(dag_state);
-        let mut dag_state = DagState::new(context.clone(), store.clone());
-
-        // Since the last accepted blocks were not flushed, the equivocations from
-        // rounds 13 to 20 should not be accounted for. The metrics should remain
-        // the same as before this acceptance.
-        assert_eq!(
-            [
-                metrics.uncached_equivocations_by_authority(),
-                metrics.uncached_missing_proposals_by_authority(),
-                metrics.equivocations_in_cache_by_authority(),
-                metrics.missing_proposals_in_cache_by_authority()
-            ],
-            [
-                vec![0; committee_size],
-                vec![3, 0, 0, 0],
-                vec![0, 1, 0, 0],
-                vec![0; committee_size],
-            ]
-        );
-
-        // Now we accept those lost blocks again and flush the dag state
-        dag_state.accept_blocks(dag_builder.blocks(13..=20));
-        for commit in second_temp_commits.clone() {
-            dag_state.add_commit(commit);
-        }
-        dag_state.flush();
-
-        // Now all misbehaviours should be accounted for in the uncached metrics.
-        assert_eq!(
-            [
-                metrics.uncached_equivocations_by_authority(),
-                metrics.uncached_missing_proposals_by_authority(),
-                metrics.equivocations_in_cache_by_authority(),
-                metrics.missing_proposals_in_cache_by_authority()
-            ],
-            [
-                vec![0, 1, 2, 0],
-                vec![3, 0, 0, 0],
-                vec![0; committee_size],
-                vec![0; committee_size],
-            ]
-        );
-    }
-
-    #[tokio::test]
-    async fn test_metrics_flush_and_recovery() {
-        telemetry_subscribers::init_for_testing();
-
-        const GC_DEPTH: u32 = 0;
-        const CACHED_ROUNDS: u32 = 5;
-
-        let committee_size = 4;
-        let (mut context, _) = Context::new_for_test(committee_size);
-
-        context.parameters.dag_state_cached_rounds = CACHED_ROUNDS;
-        context
-            .protocol_config
-            .set_consensus_gc_depth_for_testing(GC_DEPTH);
-        context
-            .protocol_config
-            .set_consensus_linearize_subdag_v2_for_testing(false);
-
-        let context = Arc::new(context);
-        let metrics = &context.metrics.scoring_metrics;
-        let store = Arc::new(MemStore::new());
-        // `cached_rounds` is initialized here as 5.
-        let mut dag_state = DagState::new(context.clone(), store.clone());
-
-        // Initialize the DAG builder with 20 layers. Blocks in the DAG will reference
-        // all blocks from the previous round.
-        // - Rounds 1 to 5 will have unique blocks from all authorities.
-        // - Rounds 6 to 8 will have unique blocks from all authorities, except 0, who
-        //   will not propose any block.
-        // - Rounds 9 to 10 will have unique blocks from all authorities.
-        // - Rounds 11 to 20 will have unique blocks from all authorities, except:
-        //      - Authority 1, who will produce 1 equivocating blocks at round 11 (i.e.,
-        //        1+1 blocks)
-        //      - Authority 2, who will produce 2 equivocating blocks at round 13 (i.e.,
-        //        1+2 blocks)
-        let mut dag_builder = DagBuilder::new(context.clone());
-        dag_builder.layers(1..=5).build();
-        dag_builder
-            .layers(6..=8)
-            .authorities(vec![AuthorityIndex::new_for_test(0)])
-            .skip_block()
-            .build();
-        dag_builder.layers(9..=10).build();
-        dag_builder
-            .layers(11..=11)
-            .authorities(vec![AuthorityIndex::new_for_test(1)])
-            .equivocate(1)
-            .build();
-        dag_builder.layers(12..=12).build();
-        dag_builder
-            .layers(13..=13)
-            .authorities(vec![AuthorityIndex::new_for_test(2)])
-            .equivocate(2)
-            .build();
-        dag_builder.layers(14..=20).build();
-
-        let mut commits = dag_builder
-            .get_sub_dag_and_commits(1..=20)
-            .into_iter()
-            .map(|(_subdag, commit)| commit)
-            .collect::<Vec<_>>();
-
-        // Add the blocks and commits from first 10 rounds to the dag state. Since
-        // authority 0 skipped a leader round, we use the 9 first items of the commits
-        // vector
-        let mut temp_commits = commits.split_off(9);
-        dag_state.accept_blocks(dag_builder.blocks(1..=10));
-        for commit in commits.clone() {
-            dag_state.add_commit(commit);
-        }
-
-        // Checks that metrics are still all zeroed, since even though we accepted
-        // blocks to the dag state, the metrics updates are done when the dag state is
-        // flushed.
-        assert_eq!(
-            [
-                metrics.uncached_equivocations_by_authority(),
-                metrics.uncached_missing_proposals_by_authority(),
-                metrics.equivocations_in_cache_by_authority(),
-                metrics.missing_proposals_in_cache_by_authority()
-            ],
-            [
-                vec![0; committee_size],
-                vec![0; committee_size],
-                vec![0; committee_size],
-                vec![0; committee_size]
-            ]
-        );
-
-        // Flush the dag state
-        dag_state.flush();
-
-        // Check that metrics were updated correctly after flushing.
-        //
-        // Equivocations:
-        // - We only accepted blocks from rounds <= 10, thus, no equivocations were
-        //   accepted yet. Equivocations metrics, then, should be still all zeroed.
-        //
-        // Missing proposals:
-        // - The the last committed round is 10, so the eviction round should be 5 for
-        //   authority 2 (leader of round 10) and 4 for all other authorities.
-        // - The threshold_clock_round should be 11, since we already accepted all
-        //   blocks from epoch 10.
-        // - Then, finally, we should have counted:
-        //      - 0 uncached missing proposals for authority 0;
-        //      - 3 missing proposal in cache for authority 0;
-        //      - 0 missing proposals for authorities 1, 2, and 3.
-        assert_eq!(
-            [
-                metrics.uncached_equivocations_by_authority(),
-                metrics.uncached_missing_proposals_by_authority(),
-                metrics.equivocations_in_cache_by_authority(),
-                metrics.missing_proposals_in_cache_by_authority()
-            ],
-            [
-                vec![0; committee_size],
-                vec![0; committee_size],
-                vec![0; committee_size],
-                vec![3, 0, 0, 0],
-            ]
-        );
-
-        // Clear and check all metrics
-        metrics.uncached[0]
-            .missing_proposals
-            .store(0, Ordering::Relaxed);
-        metrics.cached[0]
-            .missing_proposals
-            .store(0, Ordering::Relaxed);
-        assert_eq!(
-            [
-                metrics.uncached_equivocations_by_authority(),
-                metrics.uncached_missing_proposals_by_authority(),
-                metrics.equivocations_in_cache_by_authority(),
-                metrics.missing_proposals_in_cache_by_authority()
-            ],
-            [
-                vec![0; committee_size],
-                vec![0; committee_size],
-                vec![0; committee_size],
-                vec![0; committee_size]
-            ]
-        );
-
-        // Destroy and recover dag state from storage.
-        drop(dag_state);
-        let mut dag_state = DagState::new(context.clone(), store.clone());
-
-        assert_eq!(dag_state.last_commit_index(), 9);
-        assert_eq!(dag_state.last_committed_rounds(), [9, 9, 10, 9]);
-
-        // Metrics should have been initialized as before the recovery.
-        assert_eq!(
-            [
-                metrics.uncached_equivocations_by_authority(),
-                metrics.uncached_missing_proposals_by_authority(),
-                metrics.equivocations_in_cache_by_authority(),
-                metrics.missing_proposals_in_cache_by_authority()
-            ],
-            [
-                vec![0; committee_size],
-                vec![0; committee_size],
-                vec![0; committee_size],
-                vec![3, 0, 0, 0],
-            ]
-        );
-
-        // Add blocks and commits from rounds 11 and 12 to the dag state.
-        let second_temp_commits = temp_commits.split_off(2);
-        dag_state.accept_blocks(dag_builder.blocks(11..=12));
-        for commit in temp_commits.clone() {
-            dag_state.add_commit(commit);
-        }
-
-        // Flush the dag state
-        dag_state.flush();
-
-        // Check that metrics were updated correctly after flushing.
-        //
-        // Missing proposals:
-        // - The last commit round is 12, so the eviction round should be 7 for
-        //   authority 0 (leader of round 12) and 6 for all other authorities. Then, we
-        //   should have counted:
-        //      - 2 uncached missing proposals for authority 0;
-        //      - 1 missing proposal in cache for authority 0;
-        //      - 0 missing proposals for authorities 1, 2, and 3.
-        //
-        // Equivocations:
-        // - We only removed from cache blocks from rounds <= 7, thus, no equivocations
-        //   should be uncached. Then, we should have counted:
-        //      - 0 uncached equivocations;
-        //      - 1 equivocation in cache for authority 1;
-        //      - 0 equivocations in cache for authorities 0, 2 and 3;
-        //
-
-        assert_eq!(
-            [
-                metrics.uncached_equivocations_by_authority(),
-                metrics.uncached_missing_proposals_by_authority(),
-                metrics.equivocations_in_cache_by_authority(),
-                metrics.missing_proposals_in_cache_by_authority()
-            ],
-            [
-                vec![0; committee_size],
-                vec![2, 0, 0, 0],
-                vec![0, 1, 0, 0],
-                vec![1, 0, 0, 0],
-            ]
-        );
-
-        // Accept all the rest of blocks and commits.
-        dag_state.accept_blocks(dag_builder.blocks(13..=20));
-        for commit in second_temp_commits.clone() {
-            dag_state.add_commit(commit);
-        }
-
-        // Clear and check all metrics
-        metrics.uncached[0]
-            .missing_proposals
-            .store(0, Ordering::Relaxed);
-        metrics.cached[1].equivocations.store(0, Ordering::Relaxed);
-        metrics.cached[0]
-            .missing_proposals
-            .store(0, Ordering::Relaxed);
-
-        assert_eq!(
-            [
-                metrics.uncached_equivocations_by_authority(),
-                metrics.uncached_missing_proposals_by_authority(),
-                metrics.equivocations_in_cache_by_authority(),
-                metrics.missing_proposals_in_cache_by_authority()
-            ],
-            [
-                vec![0; committee_size],
-                vec![0; committee_size],
-                vec![0; committee_size],
-                vec![0; committee_size]
-            ]
-        );
-
-        // Destroy and recover dag state from storage.
-        drop(dag_state);
-        let mut dag_state = DagState::new(context.clone(), store.clone());
-
-        // Since the last accepted blocks were not flushed, the equivocations from
-        // rounds 13 to 20 should not be accounted for. The metrics should remain
-        // the same as before this acceptance.
-        assert_eq!(
-            [
-                metrics.uncached_equivocations_by_authority(),
-                metrics.uncached_missing_proposals_by_authority(),
-                metrics.equivocations_in_cache_by_authority(),
-                metrics.missing_proposals_in_cache_by_authority()
-            ],
-            [
-                vec![0; committee_size],
-                vec![2, 0, 0, 0],
-                vec![0, 1, 0, 0],
-                vec![1, 0, 0, 0],
-            ]
-        );
-
-        // Now we accept those lost blocks again and flush the dag state
-        dag_state.accept_blocks(dag_builder.blocks(13..=20));
-        for commit in second_temp_commits.clone() {
-            dag_state.add_commit(commit);
-        }
-        dag_state.flush();
-
-        // Now all misbehaviours should be accounted for in the uncached metrics.
-        assert_eq!(
-            [
-                metrics.uncached_equivocations_by_authority(),
-                metrics.uncached_missing_proposals_by_authority(),
-                metrics.equivocations_in_cache_by_authority(),
-                metrics.missing_proposals_in_cache_by_authority()
-            ],
-            [
-                vec![0, 1, 2, 0],
-                vec![3, 0, 0, 0],
-                vec![0; committee_size],
-                vec![0; committee_size],
-            ]
-        );
-    }
-
-    #[tokio::test]
-    async fn test_metrics_handle_send_block() {
-        // Initialize context and authority service given a committee_size
-        let committee_size = 4;
-        let (_, context, _, _) = new_authority_service_for_metrics_tests(committee_size);
-        let metrics = &context.metrics.scoring_metrics;
-        let source = "handle_send_block";
-        let hostname = "hostname";
-        // Create a set of errors to test
-        let ignored_error = ConsensusError::Shutdown;
-        let parsing_error = ConsensusError::MalformedBlock(bcs::Error::Eof);
-        let block_verification_error = ConsensusError::BlockRejected {
-            block_ref: BlockRef::new(10, AuthorityIndex::new_for_test(10), BlockDigest::MIN),
-            reason: "string".to_string(),
-        };
-
-        let authorities = (0..committee_size as u32)
-            .map(AuthorityIndex::new_for_test)
-            .collect::<Vec<_>>();
-
-        // Update metrics for each authority with an error that should be ignored.
-        // Metrics should not be updated for this error.
-        for authority in authorities.iter() {
-            context.metrics.update_scoring_metrics_on_block_receival(
-                *authority,
-                hostname,
-                ignored_error.clone(),
-                source,
-            );
-        }
-        assert_eq!(
-            [
-                metrics.faulty_blocks_provable_by_authority(),
-                metrics.faulty_blocks_unprovable_by_authority()
-            ],
-            [vec![0, 0, 0, 0], vec![0, 0, 0, 0]]
-        );
-
-        // Update metrics for each authority with a parsing error.
-        // Only unprovable metrics should be updated for this error.
-        for authority in authorities.iter() {
-            context.metrics.update_scoring_metrics_on_block_receival(
-                *authority,
-                hostname,
-                parsing_error.clone(),
-                source,
-            );
-        }
-        assert_eq!(
-            [
-                metrics.faulty_blocks_provable_by_authority(),
-                metrics.faulty_blocks_unprovable_by_authority()
-            ],
-            [vec![0, 0, 0, 0], vec![1, 1, 1, 1]]
-        );
-
-        // Update metrics for each authority with a signed block verification error.
-        // Only provable metrics should be updated for this error.
-        for authority in authorities.iter() {
-            context.metrics.update_scoring_metrics_on_block_receival(
-                *authority,
-                hostname,
-                block_verification_error.clone(),
-                source,
-            );
-        }
-        assert_eq!(
-            [
-                metrics.faulty_blocks_provable_by_authority(),
-                metrics.faulty_blocks_unprovable_by_authority()
-            ],
-            [vec![1, 1, 1, 1], vec![1, 1, 1, 1]]
-        );
-    }
-
-    #[tokio::test]
-    async fn test_metrics_fetch_once() {
-        // Initialize context and authority service given a committee_size
-        let committee_size = 4;
-        let (_, context, _, _) = new_authority_service_for_metrics_tests(committee_size);
-        let metrics = &context.metrics.scoring_metrics;
-        let source = "fetch_once";
-        let hostname = "hostname";
-        // Create a set of errors to test
-        let ignored_error = ConsensusError::Shutdown;
-        let parsing_error = ConsensusError::MalformedBlock(bcs::Error::Eof);
-        let block_verification_error = ConsensusError::TooManyAncestors(2, 2);
-
-        let authorities = (0..committee_size as u32)
-            .map(AuthorityIndex::new_for_test)
-            .collect::<Vec<_>>();
-
-        // Update metrics for each authority with an error that should be ignored.
-        // Metrics should not be updated for this error.
-        for authority in authorities.iter() {
-            context.metrics.update_scoring_metrics_on_block_receival(
-                *authority,
-                hostname,
-                ignored_error.clone(),
-                source,
-            );
-        }
-        assert_eq!(
-            [
-                metrics.faulty_blocks_provable_by_authority(),
-                metrics.faulty_blocks_unprovable_by_authority()
-            ],
-            [vec![0, 0, 0, 0], vec![0, 0, 0, 0]]
-        );
-
-        // Update metrics for each authority with a parsing error.
-        // Only unprovable metrics should be updated for this error.
-        for authority in authorities.iter() {
-            context.metrics.update_scoring_metrics_on_block_receival(
-                *authority,
-                hostname,
-                parsing_error.clone(),
-                source,
-            );
-        }
-        assert_eq!(
-            [
-                metrics.faulty_blocks_provable_by_authority(),
-                metrics.faulty_blocks_unprovable_by_authority()
-            ],
-            [vec![0, 0, 0, 0], vec![1, 1, 1, 1]]
-        );
-
-        // Update metrics for each authority with a signed block verification error.
-        // Since for error comes from the commit syncer, blocks received are not
-        // necessarily from the peer. Thus, it is not provable that the peer actually
-        // sent this block. Only unprovable metrics should be updated for this error.
-        for authority in authorities.iter() {
-            context.metrics.update_scoring_metrics_on_block_receival(
-                *authority,
-                hostname,
-                block_verification_error.clone(),
-                source,
-            );
-        }
-        assert_eq!(
-            [
-                metrics.faulty_blocks_provable_by_authority(),
-                metrics.faulty_blocks_unprovable_by_authority()
-            ],
-            [vec![0, 0, 0, 0], vec![2, 2, 2, 2]]
-        );
-    }
-
-    #[tokio::test]
-    async fn test_metrics_process_fetched_blocks() {
-        // Initialize context and authority service given a committee_size
-        let committee_size = 4;
-        let (_, context, _, _) = new_authority_service_for_metrics_tests(committee_size);
-        let metrics = &context.metrics.scoring_metrics;
-        let source = "process_fetched_blocks";
-        let hostname = "hostname";
-        // Create a set of errors to test
-        let ignored_error = ConsensusError::Shutdown;
-        let parsing_error = ConsensusError::MalformedBlock(bcs::Error::Eof);
-        let block_verification_error = ConsensusError::TooManyAncestors(2, 2);
-
-        let authorities = (0..committee_size as u32)
-            .map(AuthorityIndex::new_for_test)
-            .collect::<Vec<_>>();
-
-        // Update metrics for each authority with an error that should be ignored.
-        // Metrics should not be updated for this error.
-        for authority in authorities.iter() {
-            context.metrics.update_scoring_metrics_on_block_receival(
-                *authority,
-                hostname,
-                ignored_error.clone(),
-                source,
-            );
-        }
-        assert_eq!(
-            [
-                metrics.faulty_blocks_provable_by_authority(),
-                metrics.faulty_blocks_unprovable_by_authority()
-            ],
-            [vec![0, 0, 0, 0], vec![0, 0, 0, 0]]
-        );
-
-        // Update metrics for each authority with a parsing error.
-        // Only unprovable metrics should be updated for this error.
-        for authority in authorities.iter() {
-            context.metrics.update_scoring_metrics_on_block_receival(
-                *authority,
-                hostname,
-                parsing_error.clone(),
-                source,
-            );
-        }
-        assert_eq!(
-            [
-                metrics.faulty_blocks_provable_by_authority(),
-                metrics.faulty_blocks_unprovable_by_authority()
-            ],
-            [vec![0, 0, 0, 0], vec![1, 1, 1, 1]]
-        );
-
-        // Update metrics for each authority with a signed block verification error.
-        // Since for error comes from the synchronizer, blocks received are not
-        // necessarily from the peer. Thus, it is not provable that the peer actually
-        // sent this block. Only unprovable metrics should be updated for this error.
-        for authority in authorities.iter() {
-            context.metrics.update_scoring_metrics_on_block_receival(
-                *authority,
-                hostname,
-                block_verification_error.clone(),
-                source,
-            );
-        }
-        assert_eq!(
-            [
-                metrics.faulty_blocks_provable_by_authority(),
-                metrics.faulty_blocks_unprovable_by_authority()
-            ],
-            [vec![0, 0, 0, 0], vec![2, 2, 2, 2]]
-        );
-    }
+pub(crate) fn test_metrics() -> Arc<Metrics> {
+    initialise_metrics(Registry::new())
 }
