+++ consensus/core/src/synchronizer.rs
@@ -520,11 +520,12 @@ impl<C: NetworkClient, V: BlockVerifier, D: CoreThreadDispatcher> Synchronizer<C
                                 commands_sender.clone(),
                                 "live"
                             ).await {
-                                context.metrics.update_scoring_metrics_on_block_receival(
+                                context.scorer.update_scoring_metrics_on_block_receival(
                                     peer_index,
                                     peer_hostname,
                                     err.clone(),
                                     "process_fetched_blocks",
+                                    &context.metrics.node_metrics,
                                 );
                                 warn!("Error while processing fetched blocks from peer {peer_index} {peer_hostname}: {err}");
                                 context.metrics.node_metrics.synchronizer_process_fetched_failures_by_peer.with_label_values(&[peer_hostname.as_str(), "live"]).inc();
